otouto 3.13 good lord

+++ tg-launch.sh
@@ -1,11 +1,11 @@
 #!/bin/sh
 
 # Launch tg listening on the default port (change this if you've changed it in
-# config.lua), delete state file after stop, wait two seconds, and restart.
+# config.lua), delete state file after stop, wait five seconds, and restart.
 
 while true; do
 	tg/bin/telegram-cli -P 4567 -E
-	rm ~/.telegram-cli/state
+	[ -f ~/.telegram-cli/state ] && rm ~/.telegram-cli/state
 	echo 'tg has stopped. ^C to exit.'
 	sleep 5s
 done
